requirements/base.txt
# Core Django and Python dependencies
Django==4.2.11
python-dotenv==1.0.1

# Django REST Framework
djangorestframework==3.15.1
django-filter==24.2
drf-spectacular==0.27.2

# Authentication & Authorization
django-allauth==0.63.2
dj-rest-auth==6.0.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Celery & Task Queue
celery==5.3.6
redis==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1

# CORS & Security
django-cors-headers==4.3.1
django-environ==0.11.2

# File Storage & Media
Pillow==10.3.0
django-storages[google]==1.14.2

# Audit & History
django-simple-history==3.5.0

# Utilities
python-dateutil==2.9.0
pytz==2024.1
requests==2.31.0
